# api/api.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "api/api.proto",
    "version": "version not set"
  },
  "tags": [
    {
      "name": "catalog"
    }
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/v1/Good": {
      "delete": {
        "operationId": "catalog_GoodDelete",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "type": "object",
              "properties": {}
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/apiGoodDeleteRequest"
            }
          }
        ],
        "tags": [
          "catalog"
        ]
      },
      "post": {
        "operationId": "catalog_GoodCreate",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "type": "object",
              "properties": {}
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/apiGoodCreateRequest"
            }
          }
        ],
        "tags": [
          "catalog"
        ]
      },
      "put": {
        "operationId": "catalog_GoodUpdate",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "type": "object",
              "properties": {}
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/apiGoodUpdateRequest"
            }
          }
        ],
        "tags": [
          "catalog"
        ]
      }
    },
    "/v1/Good/{code}": {
      "get": {
        "operationId": "catalog_GoodGet",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/apiGoodGetResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "required": true,
            "type": "string",
            "format": "uint64"
          }
        ],
        "tags": [
          "catalog"
        ]
      }
    },
    "/v1/Goods": {
      "get": {
        "operationId": "catalog_GoodList",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/apiGoodListResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "tags": [
          "catalog"
        ]
      }
    }
  },
  "definitions": {
    "GoodListResponseGood": {
      "type": "object",
      "properties": {
        "code": {
          "type": "string",
          "format": "uint64"
        },
        "name": {
          "type": "string"
        }
      }
    },
    "apiGoodCreateRequest": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "unitOfMeasure": {
          "type": "string"
        },
        "country": {
          "type": "string"
        }
      }
    },
    "apiGoodDeleteRequest": {
      "type": "object",
      "properties": {
        "code": {
          "type": "string",
          "format": "uint64"
        }
      }
    },
    "apiGoodGetResponse": {
      "type": "object",
      "properties": {
        "code": {
          "type": "string",
          "format": "uint64"
        },
        "name": {
          "type": "string"
        },
        "unitOfMeasure": {
          "type": "string"
        },
        "country": {
          "type": "string"
        }
      }
    },
    "apiGoodListResponse": {
      "type": "object",
      "properties": {
        "Goods": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/GoodListResponseGood"
          }
        }
      }
    },
    "apiGoodUpdateRequest": {
      "type": "object",
      "properties": {
        "code": {
          "type": "string",
          "format": "uint64"
        },
        "name": {
          "type": "string"
        },
        "unitOfMeasure": {
          "type": "string"
        },
        "country": {
          "type": "string"
        }
      }
    },
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    }
  }
}